refactor(adapters): align Discord outbound extraction with Telegram pattern (#649)
Extract _make_streaming_callbacks from discord.py to discord_outbound.build_streaming_callbacks(),
add _validate_inbound() helper to discord_formatting, and align render_audio/render_attachment
signatures to take (adapter, msg, inbound) matching Telegram's convention.
