+++ src/lyra/adapters/discord_audio_outbound.py
@@ -3,7 +3,7 @@
 from __future__ import annotations
 
 import logging
-from collections.abc import AsyncIterator, Awaitable, Callable
+from collections.abc import AsyncIterator
 from io import BytesIO
 from typing import TYPE_CHECKING, Any
 
@@ -11,6 +11,7 @@ import discord
 
 from lyra.adapters._shared import (
     _AUDIO_EXTS,
+    ATTACHMENT_EXTS_BASE,
     DISCORD_MAX_LENGTH,
     buffer_and_render_audio,
     mime_to_ext,
@@ -18,6 +19,7 @@ from lyra.adapters._shared import (
     sanitize_filename,
     truncate_caption,
 )
+from lyra.adapters.discord_formatting import _validate_inbound
 from lyra.core.message import (
     InboundMessage,
     OutboundAttachment,
@@ -27,7 +29,7 @@ from lyra.core.message import (
 )
 
 if TYPE_CHECKING:
-    from lyra.adapters.discord_voice import VoiceSessionManager
+    from lyra.adapters.discord import DiscordAdapter
 
 log = logging.getLogger(__name__)
 
@@ -36,12 +38,9 @@ _VOICE_MESSAGE_FLAG = 8192
 
 
 async def render_audio(
+    adapter: "DiscordAdapter",
     msg: OutboundAudio,
     inbound: InboundMessage,
-    *,
-    bot_id: str,
-    resolve_channel: Any,
-    http: Any,
 ) -> None:
     """Send an OutboundAudio envelope as a Discord voice message.
 
@@ -49,26 +48,13 @@ async def render_audio(
     renders a proper voice bubble with waveform and inline playback.
     Falls back to regular file attachment if conversion fails.
     """
-    if inbound.platform != Platform.DISCORD.value:
-        log.error(
-            "render_audio() called with non-discord message id=%s",
-            inbound.id,
-        )
+    meta = _validate_inbound(inbound, "render_audio")
+    if meta is None:
         return
-
-    channel_id: int | None = inbound.platform_meta.get("channel_id")
-    if channel_id is None:
-        log.error(
-            "render_audio: platform_meta missing 'channel_id' for msg id=%s",
-            inbound.id,
-        )
-        return
-
-    thread_id: int | None = inbound.platform_meta.get("thread_id")
+    channel_id, thread_id, message_id = meta
     send_to_id = thread_id if thread_id is not None else channel_id
 
     # Determine message to reply to
-    message_id: int | None = inbound.platform_meta.get("message_id")
     reply_to_id = parse_reply_to_id(msg.reply_to_id)
     if reply_to_id is None and thread_id is None:
         reply_to_id = message_id
@@ -109,7 +95,7 @@ async def render_audio(
         "POST", "/channels/{channel_id}/messages", channel_id=send_to_id
     )
     try:
-        await http.request(route, form=form, files=[voice_file])
+        await adapter.http.request(route, form=form, files=[voice_file])
         log.info(
             "render_audio: voice message sent (%d bytes OGG, %.1fs) for msg id=%s",
             len(msg.audio_bytes),
@@ -121,18 +107,16 @@ async def render_audio(
             "render_audio: voice message failed — falling back to file attachment",
             exc_info=True,
         )
-        messageable = await resolve_channel(send_to_id)
+        messageable = await adapter._resolve_channel(send_to_id)
         ext = mime_to_ext(msg.mime_type, _AUDIO_EXTS)
         attachment = discord.File(fp=BytesIO(msg.audio_bytes), filename=f"audio.{ext}")
         await messageable.send(content=content or None, file=attachment)
 
 
 async def render_attachment(
+    adapter: "DiscordAdapter",
     msg: OutboundAttachment,
     inbound: InboundMessage,
-    *,
-    resolve_channel: Any,
-    attachment_exts: frozenset[str],
 ) -> None:
     """Send an OutboundAttachment envelope as a Discord file attachment.
 
@@ -140,40 +124,27 @@ async def render_attachment(
     Caption (if set) is passed as message content. Reply and thread routing
     follow the same pattern as render_audio.
     """
-    if inbound.platform != Platform.DISCORD.value:
-        log.error(
-            "render_attachment() called with non-discord message id=%s",
-            inbound.id,
-        )
+    meta = _validate_inbound(inbound, "render_attachment")
+    if meta is None:
         return
-
-    channel_id: int | None = inbound.platform_meta.get("channel_id")
-    if channel_id is None:
-        log.error(
-            "render_attachment: platform_meta missing 'channel_id' for msg id=%s",
-            inbound.id,
-        )
-        return
-
-    thread_id: int | None = inbound.platform_meta.get("thread_id")
+    channel_id, thread_id, message_id = meta
     send_to_id = thread_id if thread_id is not None else channel_id
-    messageable = await resolve_channel(send_to_id)
+    messageable = await adapter._resolve_channel(send_to_id)
 
     # Derive filename: sanitize explicit name or derive from mime_type.
     if msg.filename:
         filename = sanitize_filename(
             msg.filename,
-            attachment_exts,
+            ATTACHMENT_EXTS_BASE,
         )
     else:
-        ext = mime_to_ext(msg.mime_type, attachment_exts)
+        ext = mime_to_ext(msg.mime_type, ATTACHMENT_EXTS_BASE)
         filename = f"attachment.{ext}"
 
     buf = BytesIO(msg.data)
     file_obj = discord.File(fp=buf, filename=filename)
 
     # Determine reply target
-    message_id: int | None = inbound.platform_meta.get("message_id")
     reply_to_id = parse_reply_to_id(msg.reply_to_id)
     if reply_to_id is None and thread_id is None:
         reply_to_id = message_id
@@ -197,26 +168,29 @@ async def render_attachment(
 
 
 async def render_audio_stream(
+    adapter: "DiscordAdapter",
     chunks: AsyncIterator[OutboundAudioChunk],
     inbound: InboundMessage,
-    render_audio_fn: Callable[[OutboundAudio, InboundMessage], Awaitable[None]],
 ) -> None:
     """Buffer streamed audio chunks and send as a single Discord file attachment."""
-    if inbound.platform != Platform.DISCORD.value:
-        log.error(
-            "render_audio_stream() called with non-discord message id=%s",
-            inbound.id,
-        )
+    meta = _validate_inbound(inbound, "render_audio_stream")
+    if meta is None:
         return
-    await buffer_and_render_audio(chunks, inbound, render_audio_fn)
+    await buffer_and_render_audio(
+        chunks, inbound, lambda audio, msg: render_audio(adapter, audio, msg)
+    )
 
 
 async def render_voice_stream(
+    adapter: "DiscordAdapter",
     chunks: AsyncIterator[OutboundAudioChunk],
     inbound: InboundMessage,
-    vsm: "VoiceSessionManager",
 ) -> None:
-    """Route TTS stream to the active Discord voice session for this guild."""
+    """Route TTS stream to the active Discord voice session for this guild.
+
+    Uses platform-only check (not _validate_inbound) because voice streams
+    route by guild_id, not channel_id.
+    """
     if inbound.platform != Platform.DISCORD.value:
         log.warning(
             "render_voice_stream() called with non-discord message id=%s",
@@ -230,4 +204,4 @@ async def render_voice_stream(
             inbound.id,
         )
         return
-    await vsm.stream(str(guild_id), chunks)
+    await adapter._vsm.stream(str(guild_id), chunks)
